requirements.txt
Flask
gunicorn
google-cloud-vision
google-cloud-storage
google-cloud-datastore
